# services/user/sdk/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/user/{userIdentity}/job": {
            "get": {
                "operationId": "JobController_get",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobsViewResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "put": {
                "operationId": "JobController_insert",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateJobBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateJobResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/job/{identity}": {
            "delete": {
                "operationId": "JobController_delete",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    },
                    {
                        "name": "identity",
                        "required": true,
                        "in": "path",
                        "description": "Identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/bankAccount": {
            "get": {
                "operationId": "BankAccountController_get",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BankAccountsViewResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "put": {
                "operationId": "BankAccountController_insert",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBankAccountBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateBankAccountResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/bankAccount/{identity}": {
            "delete": {
                "operationId": "BankAccountController_delete",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    },
                    {
                        "name": "identity",
                        "required": true,
                        "in": "path",
                        "description": "Identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Finance API",
        "description": "Finance API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "scopes": {
                            "openid": "openid",
                            "profile": "profile",
                            "email": "email"
                        },
                        "authorizationUrl": "https://dev-tzti25ivp5zmox2s.us.auth0.com/authorize"
                    }
                },
                "in": "header",
                "name": "Authorization",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "PaginatedResponsePage": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "number",
                        "description": "The amount available in this response",
                        "example": 10
                    },
                    "offset": {
                        "type": "number",
                        "description": "The starting index of the paginated data",
                        "example": 0
                    },
                    "total": {
                        "type": "number",
                        "description": "The total number of items available",
                        "example": 10
                    }
                },
                "required": [
                    "count",
                    "offset",
                    "total"
                ]
            },
...
            "HttpErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "integer",
                        "description": "The http code returned by the server"
                    },
                    "message": {
                        "type": "string",
                        "description": "A message describing the error that occurred"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "BalanceResponse": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "currency"
                ]
            },
            "BankAccountResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "bank": {
                        "type": "string"
                    },
                    "balance": {
                        "$ref": "#/components/schemas/BalanceResponse"
                    }
                },
                "required": [
                    "identity",
                    "bank",
                    "balance"
                ]
            },
            "ActiveIncomeResponse": {
                "type": "object",
                "properties": {
                    "monthlySalary": {
                        "type": "number"
                    }
                },
                "required": [
                    "monthlySalary"
                ]
            },
            "JobResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "activeIncome": {
                        "$ref": "#/components/schemas/ActiveIncomeResponse"
                    }
                },
                "required": [
                    "identity",
                    "title",
                    "activeIncome"
                ]
            },
...
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "PaginatedJobResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "$ref": "#/components/schemas/PaginatedResponsePage"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JobResponse"
                        }
                    }
                },
                "required": [
                    "page",
                    "data"
                ]
            },
            "JobsViewResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedJobResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "CreateJobBody": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "monthlySalary": {
                        "type": "number",
                        "minimum": 0
                    }
                },
                "required": [
                    "title",
                    "monthlySalary"
                ]
            },
            "CreateJobResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/JobResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "PaginatedBankAccountResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "$ref": "#/components/schemas/PaginatedResponsePage"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BankAccountResponse"
                        }
                    }
                },
                "required": [
                    "page",
                    "data"
                ]
            },
            "BankAccountsViewResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedBankAccountResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "CreateBankAccountBody": {
                "type": "object",
                "properties": {
                    "bank": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "number",
                        "minimum": 0
                    },
                    "currency": {
                        "type": "string"
                    }
                },
                "required": [
                    "bank",
                    "balance",
                    "currency"
                ]
            },
            "CreateBankAccountResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/BankAccountResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
        }
    }
}